shioriProject/requirements.txt
flask-sqlalchemy==2.4.0
mysql-python==1.2.5
flask-login==0.4.1
flask-migrate==2.5.2
Flask-WTF==0.14.2
flask-bootstrap==3.3.7.1
flask-mail==0.9.1
pyjwt==1.7.1
flask-bcrypt==0.7.1
